% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsia.R
\name{calculate_G}
\alias{calculate_G}
\title{Correlation matrix G}
\usage{
calculate_G(g)
}
\arguments{
\item{g}{undirected graph}
}
\value{
numeric matrix G
}
\description{
Correlation matrix G
}
\examples{
df0<-as.data.frame(matrix(c(1,2,2,3,3,4,4,5,5,6,4,7,5,7,3,7,7,6,3,8,8,9,6,10,1,11,2,7,11,13,11,14,11,15,11,16,11,17,11,18,11,19,11,20,19,20,20,21,2,22,6,2,17,8,9,23,23,24,12,24,23,25,25,26,25,27,27,28),ncol=2,byrow=TRUE))
species<-as.character(1:28)
g<-graph_from_data_frame(df0,directed=FALSE,vertices=species)
set.seed(135)
plot(g,vertex.size=11)
G<-calculate_G(g)
}
\seealso{
""
}
