% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsia.R
\name{calculate_Iij}
\alias{calculate_Iij}
\title{Pairwise mutual information matrix I_{ij}}
\usage{
calculate_Iij(G)
}
\arguments{
\item{G}{correlation matrix G}
}
\value{
numeric matrix I_{ij}
}
\description{
Pairwise mutual information matrix I_{ij}
}
\examples{
Iij<-calculate_Iij(G)
}
\seealso{
""
}
