% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRING.R
\docType{data}
\name{g_STRING}
\alias{g_STRING}
\title{STRING interaction network in a graph}
\format{graph g_STRING}
\source{
\href{https://string-db.org/}{QTL Archive}
}
\usage{
data(g_STRING)
}
\description{
STRING interaction network in a graph
}
\examples{
data(g_STRING)
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{}{PubMed})
}
\keyword{datasets}
