% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsia.R
\name{get_p_tI}
\alias{get_p_tI}
\title{p-value of tI(A)}
\usage{
get_p_tI(A, G, null.domain = colnames(G), null.prob = rep(1,
  length(null.domain)), nsamples = 1000, plot = TRUE)
}
\arguments{
\item{A}{character vector of genes in set A}

\item{G}{correlation matrix G}

\item{null.domain}{gene set from which null gene sets are drawn (default colnames(G))}

\item{null.prob}{genes will be drawn with probability proportional to these weights (default is uniform probabilities).}

\item{nsamples}{number of null sets to be generated (default 1000)}

\item{plot}{logical. Should the null distribution be plotted (default TRUE)}
}
\value{
The probability that tI(R) >= tI(A), where R is a null gene set of size |A|
drawn randomly from null.domain.
Each gene is drawn without replacement with a probability proportional to null.prob.
A total of nsamples will be drawn.
}
\description{
p-value of tI(A)
}
\examples{
df0<-as.data.frame(matrix(c(1,2,2,3,3,4,4,5,5,6,4,7,5,7,3,7,7,6,3,8,8,9,6,10,1,11,2,7,11,13,11,14,11,15,11,16,11,17,11,18,11,19,11,20,19,20,20,21,2,22,6,2,17,8,9,23,23,24,12,24,23,25,25,26,25,27,27,28,29,30),ncol=2,byrow=TRUE))
vnames<-as.character(1:30)
g<-graph_from_data_frame(df0,directed=FALSE,vertices=vnames)
set.seed(1357)
plot(g,vertex.size=11)
G<-get_corr_matrix(g)
A<-c("2","4","6","30")
p<-get_p_tI(A,G)
}
\seealso{
\code{\link{get_p_I}}
}
