% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsia.R
\name{get_STRING}
\alias{get_STRING}
\title{STRING PPI network annotated with Entrez gene IDs}
\usage{
get_STRING(species = 9606, score = 700)
}
\arguments{
\item{species}{taxonomy ID}

\item{score}{minimum interaction score}
}
\value{
graph
}
\description{
STRING PPI network annotated with Entrez gene IDs
}
\examples{
#for human ppi graph with interaction score >= 700
g_STRING<-get_STRING() #for human ppi graph with interaction score >= 700

#for mouse ppi graph with score >= 500
g_STRING_mouse<-get_STRING(species=10090,score=500) #for mouse ppi graph with score >= 500
}
\seealso{
""
}
