% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsia.R
\name{I}
\alias{I}
\title{Mutual information I(kA,kB) of gene sets A and B}
\usage{
I(kA, kB, G)
}
\arguments{
\item{kA}{numeric vector of gene indices in set A}

\item{kB}{numeric vector of gene indices in set B}

\item{G}{correlation matrix G}
}
\value{
numeric I(kA,kB)
}
\description{
Mutual information I(kA,kB) of gene sets A and B
}
\examples{
indx<-1:length(V(g))
names(indx)<-as.character(V(g))
A<-c("1","17","8") #vector of gene names in set A
kA<-indx[A] #gene indices of set A
B<-c("1","11") #vector of gene names in set B
kB<-indx[B] #gene indices of set B
I(kA,kB,G)
}
\seealso{
""
}
