% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsia.R
\name{inherit_interactions}
\alias{inherit_interactions}
\title{Graph of expressed.genes as verices with interactions inherited from graph g}
\usage{
inherit_interactions(g, expressed.genes)
}
\arguments{
\item{g}{graph}

\item{expressed.genes}{character vector of expressed gene IDs}
}
\value{
graph containing all expressed.genes as vertices with interactions inherited from graph g. If any expressed.genes are not found in g, they will become unlinked vertices
}
\description{
Graph of expressed.genes as verices with interactions inherited from graph g
}
\examples{
df0<-as.data.frame(matrix(c(1,2,2,3,3,4,4,5,5,6,4,7,5,7,3,7,7,6,3,8,8,9,6,10,1,11,2,7,11,13,11,14,11,15,11,16,11,17,11,18,11,19,11,20,19,20,20,21,2,22,6,2,17,8,9,23,23,24,12,24,23,25,25,26,25,27,27,28,29,30),ncol=2,byrow=TRUE))
vnames<-as.character(1:30)
g<-graph_from_data_frame(df0,directed=FALSE,vertices=vnames)
set.seed(1357)
plot(g,vertex.size=11)
expressed.genes<-c("2","4","6","30","40")
g.expressed<-inherit_interactions(g,expressed.genes)
plot(g.expressed,vertex.size=11)
}
\seealso{
""
}
