% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsia.R
\name{tI}
\alias{tI}
\title{Interaction information I(kA) of gene set A}
\usage{
tI(kA, G)
}
\arguments{
\item{kA}{numeric vector of gene indices in set A}

\item{G}{correlation matrix G}
}
\value{
numeric I(kA)
}
\description{
Interaction information I(kA) of gene set A
}
\examples{
#here we are assuming that matrix G and graph g exist
A<-c(1,17,8) #vector of gene indices in set A
names(V(g))[A] # these are the corresponding gene names
tI(kA,G)
}
\seealso{
""
}
