% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsia.R
\name{IQR}
\alias{IQR}
\title{Information quality ratio IQR(kA,kB) of a gene set A and B}
\usage{
IQR(kA, kB, G)
}
\arguments{
\item{kA}{numeric vector of gene indices in set A}

\item{kB}{numeric vector of gene indices in set B}

\item{G}{correlation matrix G}
}
\value{
numeric IQR(kA,kB)
}
\description{
Information quality ratio IQR(kA,kB) of a gene set A and B
}
\examples{
indx<-1:length(V(g))
names(indx)<-as.character(V(g))
A<-c("1","17","8") #vector of gene names in set A
kA<-indx[A] #gene indices in set A
B<-c("1","11") #vector of gene names in set B
kB<-indx[B] #gene indices in set B
IQR(kA,kB,G)
}
\seealso{
""
}
